pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatchling",
]

[project]
name = "django-json-schema-editor"
description = "Django widget for using @json-editor/json-editor in the admin"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Matthias Kestenholz", email = "mk@feinheit.ch" },
  { name = "Fabian Germann", email = "fg@feinheit.ch" },
]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dynamic = [
  "version",
]
dependencies = [
  "django-js-asset>=3",
  "fastjsonschema>=2.19",
]
optional-dependencies.prose = [
  "django-prose-editor>=0.11a11",
]
optional-dependencies.tests = [
  "asgiref",
  "coverage",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-django",
  "pytest-playwright",
]

urls.Documentation = "https://github.com/matthiask/django-json-schema-editor#readme"
urls.Issues = "https://github.com/matthiask/django-json-schema-editor/issues"
urls.Source = "https://github.com/matthiask/django-json-schema-editor/"

[tool.hatch.version]
path = "django_json_schema_editor/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
  "/django_json_schema_editor",
]

[tool.ruff]
target-version = "py39"

fix = true
show-fixes = true
lint.extend-select = [
  # flake8-bugbear
  "B",
  # flake8-comprehensions
  "C4",
  # mmcabe
  "C90",
  # flake8-django
  "DJ",
  "E",
  # pyflakes, pycodestyle
  "F",
  # flake8-boolean-trap
  "FBT",
  # isort
  "I",
  # flake8-gettext
  "INT",
  # flake8-implicit-string-concatenation
  "ISC",
  # pep8-naming
  "N",
  # pygrep-hooks
  "PGH",
  # flake8-pie
  "PIE",
  # pylint
  "PLC",
  "PLE",
  "PLW",
  # unused noqa
  "RUF100",
  # pyupgrade
  "UP",
  "W",
  # flake8-2020
  "YTT",
]
lint.extend-ignore = [
  # Allow zip() without strict=
  "B905",
  # No line length errors
  "E501",
]
lint.per-file-ignores."*/migrat*/*" = [
  # Allow using PascalCase model names in migrations
  "N806",
  # Ignore the fact that migration files are invalid module names
  "N999",
]
lint.isort.combine-as-imports = true
lint.isort.lines-after-imports = 2
lint.mccabe.max-complexity = 15

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "testapp.settings"
python_files = "test_*.py"
addopts = "--strict-markers"
testpaths = [ "tests" ]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
  "e2e: End-to-end browser tests",
]
